Add Send + Sync error bounds of firehose example main()
compiler throws an issue related to knowing size of the error at
compile time to use `?`, so either the Send + Sync markers need to go
from the process _message() function or be added to main().

At least I think?

I decided to add them to main() for no particular reason.